This is the initial commit of the project
This project is meant to become a Trainer for the PC version of the FrostHaven boardgame.
This is the first commit, which contains the base logic, a (somewhat) detailed plan for the strategy to design the program, as well as instructions on how the code is structured.

I will make this the official start of the project by pushing this to the main branch, and get the project on the road.