First working version
Can alter cards and hp, to some extent.
I had to change rhe rulesetmanager to use a different method than previously for handling the filler bank and character parsing.
I will need to test some more on this version, but also try to implement some of the new settings to the previous version of the fillerbank strategy.