Item editing now works
There is still a lot of work to be done on this front, but the basic functionality is in place.
You can now edit damage/heal/costs/shield/retaliate/conditions/usage types/range/targets etc.
Item cards are not updated in game however, so old values still show in the UI.
This is something I will need to investigate whether or not it is possible for me to fix.
I suspect I will need to access the .YAML files shown in the header/footer in all of the binary blobs.